Use JQuery for room ready status
Script can be in HTML <head> section because it is wrapped in
$(document).ready(). This will wait to run its function argument until the
document is fully loaded.

Using JQuery for the AJAX request removes several lines of code and makes it
much easier to see what is going on.

Passing the room status data as JSON in the AJAX response will make it easy to
organize any other status information we want to pass.