Refactoring through game creation
Using channels instead of polling. Also refactoring the code to make it easier
to read and use. One big feature change is that the game creator now chooses
who gets to play. Things seem to be working for game creation. The next thing
to do is to make sure the game itself is working.