Version that mostly works
I have done a few quick tests with this version and it seems to do most of the
right things. It can go all the way through the first round and start on the
second round.

Right now the biggest problem is that it doesn't know when the game is over,
and it doesn't know about the assassin.