Remove sleep before requesting info
Since the socket.onopen function was not working on the real servers, I had
switched to a sleep-based solution I found on stackoverflow. This worked, but
different devices seemed to need different sleep times and I didn't want to
choose the longest one because it created annoying lag.

Now I am using the channel presence detection URI to keep track of when the
connection is established and to notify the client over the channel. This is a
bit ugly because then the client then just immediately sends back a request for
more information. It would be nicer for the server to send the information with
the first channel message, but the server only has the client_id at that time,
and I'm not confident I can parse that to get the room name out. That would be
the best solution if I could get it to work, but I'm putting it on the back
burner for now.

I also had to put back the post calls in the socket.onopen handlers because the
presence detection does not seem to work so well on the local development
server. So basically I'm just trying to notify the client every way that I can
think of, and I'm assuming that at least one message will get through.